+++ Tests/Runtime/CompilerFramework/PrebuiltDelegates.cs
@@ -44339,25 +44339,25 @@ namespace Test.Mosa.Runtime.CompilerFramework
 		delegate int I4_I4_I4_I4_I4(int value1, int value2, int value3, int value4);
 		delegate bool B_U4_U8_U8_U8(uint value1, ulong value2, ulong value3, ulong value4);
 
-		delegate bool B_I1_I1_I1_I1(int value1, int value2, int value3, int value4);
-		delegate bool B_I2_I2_I2_I2(int value1, int value2, int value3, int value4);
+		delegate bool B_I1_I1_I1_I1(sbyte value1, sbyte value2, sbyte value3, sbyte value4);
+		delegate bool B_I2_I2_I2_I2(short value1, short value2, short value3, short value4);
 		delegate bool B_I4_I4_I4_I4(int value1, int value2, int value3, int value4);
-		delegate bool B_I8_I8_I8_I8(int value1, int value2, int value3, int value4);
+		delegate bool B_I8_I8_I8_I8(long value1, long value2, long value3, long value4);
 
-		delegate bool B_U1_U1_U1_U1(uint value1, uint value2, uint value3, uint value4);
-		delegate bool B_U2_U2_U2_U2(uint value1, uint value2, uint value3, uint value4);
+		delegate bool B_U1_U1_U1_U1(byte value1, byte value2, byte value3, byte value4);
+		delegate bool B_U2_U2_U2_U2(ushort value1, ushort value2, ushort value3, ushort value4);
 		delegate bool B_U4_U4_U4_U4(uint value1, uint value2, uint value3, uint value4);
-		delegate bool B_U8_U8_U8_U8(uint value1, uint value2, uint value3, uint value4);
+		delegate bool B_U8_U8_U8_U8(ulong value1, ulong value2, ulong value3, ulong value4);
 
-		delegate bool B_I1_I1_I1_I1_I1(int value1, int value2, int value3, int value4, int value5);
-		delegate bool B_I2_I2_I2_I2_I2(int value1, int value2, int value3, int value4, int value5);
+		delegate bool B_I1_I1_I1_I1_I1(sbyte value1, sbyte value2, sbyte value3, sbyte value4, sbyte value5);
+		delegate bool B_I2_I2_I2_I2_I2(short value1, short value2, short value3, short value4, short value5);
 		delegate bool B_I4_I4_I4_I4_I4(int value1, int value2, int value3, int value4, int value5);
-		delegate bool B_I8_I8_I8_I8_I8(int value1, int value2, int value3, int value4, int value5);
+		delegate bool B_I8_I8_I8_I8_I8(long value1, long value2, long value3, long value4, long value5);
 
-		delegate bool B_U1_U1_U1_U1_U1(uint value1, uint value2, uint value3, uint value4, uint value5);
-		delegate bool B_U2_U2_U2_U2_U2(uint value1, uint value2, uint value3, uint value4, uint value5);
+		delegate bool B_U1_U1_U1_U1_U1(byte value1, byte value2, byte value3, byte value4, byte value5);
+		delegate bool B_U2_U2_U2_U2_U2(ushort value1, short value2, short value3, short value4, short value5);
 		delegate bool B_U4_U4_U4_U4_U4(uint value1, uint value2, uint value3, uint value4, uint value5);
-		delegate bool B_U8_U8_U8_U8_U8(uint value1, uint value2, uint value3, uint value4, uint value5);
+		delegate bool B_U8_U8_U8_U8_U8(ulong value1, ulong value2, ulong value3, ulong value4, ulong value5);
 
 		delegate void V_I2_I2(short value1, short value2);
 		delegate void V_I4_I4(int value1, int value2);
